# swagger.json
{
  "swagger": "2.0",
  "basePath": "/",
  "info": {
    "title": "xCampaign",
    "description": "",
    "version": "1.0"
  },
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/auth/signup": {
      "post": {
        "x-swagger-router-controller": "home",
        "operationId": "index1",
        "tags": [
          "auth"
        ],
        "description": "[Login 123](https://www.google.com)",
        "parameters": [
          {
            "name": "name",
            "in": "formData",
            "type": "string",
            "required": "true",
            "description": "Name of Customer to create"
          },
          {
            "name": "password",
            "in": "formData",
            "type": "string",
            "required": "true",
            "description": "password to be create"
          },
          {
            "name": "email",
            "in": "formData",
            "type": "string",
            "required": "true",
            "description": "Email to be created"
          }
        ],
        "responses": {}
      }
    },
    "/v1/auth/signin": {
      "post": {
        "x-swagger-router-controller": "home",
        "operationId": "index2",
        "tags": [
          "auth"
        ],
        "description": "[Login 123](https://www.google.com)",
        "parameters": [
          {
            "name": "password",
            "in": "formData",
            "type": "string",
            "required": "true",
            "description": "password to be create"
          },
          {
            "name": "email",
            "in": "formData",
            "type": "string",
            "required": "true",
            "description": "Email to be created"
          }
        ],
        "responses": {}
      }
    },
    "/v1/community/create": {
      "post": {
        "x-swagger-router-controller": "home",
        "operationId": "index2",
        "tags": [
          "community"
        ],
        "description": "[Login 123](https://www.google.com)",
        "parameters": [
          {
            "name": "name",
            "in": "formData",
            "type": "string",
            "required": "true",
            "description": "name to be create"
          },
          {
            "name": "slug",
            "in": "formData",
            "type": "string",
            "required": "true",
            "description": "slug to be created"
          }
        ],
        "responses": {}
      }
    },
    "/add-saas": {
      "post": {
        "x-swagger-router-controller": "add-saas",
        "operationId": "index1",
        "tags": [
          "Saas Related Apis"
        ],
        "description": "This is form to add saas application",
        "parameters": [
          {
            "name": "title",
            "in": "formData",
            "type": "string",
            "required": "true",
            "description": "Name of Saas/MarketPlace to create"
          },
          {
            "name": "info",
            "in": "formData",
            "type": "string",
            "required": "true",
            "description": "Info/Description of Saas/MarketPlace to create"
          },
          {
            "name": "logoLink",
            "in": "formData",
            "type": "string",
            "required": "true",
            "description": "Link to logo of Saas/MarketPlace to create"
          },
          {
            "name": "tags",
            "in": "formData",
            "type": "string",
            "required": "true",
            "description": "Tags of Saas/MarketPlace to create"
          }
        ],
        "responses": {}
      }
    },
    "/update-saas": {
      "post": {
        "x-swagger-router-controller": "update-saas",
        "operationId": "index2",
        "tags": [
          "Saas Related Apis"
        ],
        "description": "This is form to add saas application",
        "parameters": [
          {
            "name": "id",
            "in": "formData",
            "type": "string",
            "required": "true",
            "description": "Id of Saas/MarketPlace to update"
          },
          {
            "name": "name",
            "in": "formData",
            "type": "string",
            "required": "true",
            "description": "Name of Saas/MarketPlace to update"
          },
          {
            "name": "logoLink",
            "in": "formData",
            "type": "string",
            "required": "true",
            "description": "Link to logo of Saas/MarketPlace to update"
          },
          {
            "name": "type2",
            "in": "formData",
            "type": "string",
            "required": "true",
            "description": "Tags of Saas/MarketPlace to update"
          },
          {
            "name": "info",
            "in": "formData",
            "type": "string",
            "required": "true",
            "description": "Info/Description of Saas/MarketPlace to update"
          }
        ],
        "responses": {}
      }
    }
  }
}